CVSL-868 move toJson utils method to DateTime class
